Created Student Middleware and we have done more restrictions

+++ AcademUA/app/Http/Controllers/CoursesController.php
@@ -10,41 +10,42 @@ use Illuminate\Support\Facades\Auth;
 class CoursesController extends Controller
 {
 	
+	
 	/*#############################################
-				GETTERS AND SETTERS
-	###############################################*/
-
-	public function getCourses($teacher_id){
+	GETTERS AND SETTERS
+		###############################################*/
+	
+	public function getCourses(){
 		$course = new Course();
-		$list_Courses = $course->getCourses($teacher_id);
+		$list_Courses = $course->getCourses(Auth::user()->id);
 		return view('courses.manageCourses')->with('courses', $list_Courses);
 		
 	}
-
+	
 	public function getComments($course_id){
 		$course = new Course();
 		$list_comments = $course->getComments($course_id);
 		return $list_comments;
 	}
-		
-
-		public function getCategories($course_id){
+	
+	
+	public function getCategories($course_id){
 		$course = new Course();
 		$list_categories = $course->getCategories($course_id);
 		return $list_categories;
 	}
-
-	//##############################################
+	
+	//#	#############################################
 	
 	public function edit(Request $request, $id) {
-
-
-
+		
+		
+		
 		$course = Course::findOrFail($id);
 		$this->validate($request,[
-
-				'price' => 'min:0 | numeric'
-		]);
+		
+		'price' => 'min:0 | numeric'
+				]);
 		
 		$name= $request->input('name');
 		$description= $request->input('description');
@@ -52,21 +53,31 @@ class CoursesController extends Controller
 		$content= $request->input('content');
 		$links= $request->input('links');
 		$teacher_id= $course->teacher_id;
-
+		
 		
 		$course->createCourse($name,$description,$price,$content,$links,$teacher_id);
-
-		$comments = $course->getComments($id); //returns an array with all the comments
-		return view('courses.course', ['comments' => $comments])->with('course', $course);
-
+		
+		$comments = $course->getComments($id);
+		//r		eturns an array with all the comments
+				return view('courses.course', ['comments' => $comments])->with('course', $course);
+		
 	}
 	
-	public function deleteCourse($id){ //We have to redirect to Manage Courses but we need the session of the teacher(in progress)
+	public function deleteCourse($id){
+		//We have to redirect to Manage Courses but we need the session of the teacher(in progress)
+		
 		$course = Course::findOrFail($id);
-		$course->deleteCourse();
 
-		$list = Course::paginate(6);
-		return view('courses.courses', ['courses' => $list]); //We have to change that in the future
+		if($course->teacher_id == Auth::user()->id){
+			$course->deleteCourse();
+			
+			$list = Course::paginate(6);
+			return view('courses.courses', ['courses' => $list]);
+			//We have to change that in the future
+			
+		}else{
+			return view('home');
+		}
 	}
 	
 	public function showCourses(){
@@ -74,19 +85,20 @@ class CoursesController extends Controller
 		
 		return view('courses.courses', ['courses' => $list]);
 	}
-
-	//no sabemos como pasar dos variables
-
-
+	
+	//n	o sabemos como pasar dos variables
+	
+	
 	public function showSingleCourse($id){
 		$course = Course::find($id);
-		$comments = $course->getComments($id); //returns an array with all the comments
-	return view('courses.course', ['comments' => $comments])->with('course', $course);
-
+		$comments = $course->getComments($id);
+		//r		eturns an array with all the comments
+			return view('courses.course', ['comments' => $comments])->with('course', $course);
+		
 	}
 	
-	//Mostramos cursos filtrando
-	public function showCoursesFilter(Request $request){
+	//M	ostramos cursos filtrando
+		public function showCoursesFilter(Request $request){
 		$filter = $_GET["filter"];
 		$order = $_GET["order"];
 		$how = $_GET["how"];
@@ -106,45 +118,49 @@ class CoursesController extends Controller
 	public function createCourse(Request $request){
 		$course = new Course();
 		$this->validate($request,[
-				'name' => 'required',
-				'description' => 'required',
-				'id' => 'required',
-				'price' => 'required | min:0 | numeric'
-		]);
+						'name' => 'required',
+						'description' => 'required',
+						'price' => 'required | min:0 | numeric'
+				]);
 		
 		$name= $request->input('name');
 		$description= $request->input('description');
 		$price= $request->input('price');
 		$content= $request->input('content');
 		$links= $request->input('links');
-		$teacher_id= $request->input('id');
+		$teacher_id= Auth::user()->id;
 		$categoryName = $request->input('category');
-
-
+		
+		
 		$catModel = new Category();
 		$categoryID = $catModel->getID($categoryName);
-
+		
 		
 		$course->createCourse($name,$description,$price,$content,$links,$teacher_id);
 		
 		$course->attachCategory($categoryID);
 		
-
+		
 		return view('home');
 	}
 	
 	public function modifyCourse($course_id){
 		$course = Course::find($course_id);
-		return view('/courses/modifyCourse')->with('courses', $course);
-	}
 
+		if($course->teacher_id == Auth::user()->id){
+		return view('/courses/modifyCourse')->with('courses', $course);
+		}else{
+			return view('home');
+		}
+
+	}
+	
 	
 	public function attendCourse($course_id){
 		$course = Course::find($course_id);
 		$user = User::find(Auth::user()->id);
 		$course->attendCourse($course->id, $user);
-		
 	}
-
+	
 	
 }
